--- a/src/markd/parsers/inline.cr
+++ b/src/markd/parsers/inline.cr
@@ -278,13 +278,6 @@ module Markd::Parser
     end
 
     private def process_emphasis(delimiter : Delimiter?)
-      openers_bottom = {
-        '_'  => delimiter,
-        '*'  => delimiter,
-        '\'' => delimiter,
-        '"'  => delimiter,
-      } of Char => Delimiter?
-
       # find first closer above stack_bottom:
       closer = @delimiters
       while closer
@@ -293,94 +286,103 @@ module Markd::Parser
         closer = previous
       end
 
-      # move forward, looking for closers, and handling each
-      while closer
-        closer_char = closer.char
+      if closer
+        openers_bottom = {
+          '_'  => delimiter,
+          '*'  => delimiter,
+          '\'' => delimiter,
+          '"'  => delimiter,
+        } of Char => Delimiter?
 
-        unless closer.can_close
-          closer = closer.next?
-          next
-        end
+        # move forward, looking for closers, and handling each
+        while closer
+          closer_char = closer.char
 
-        # found emphasis closer. now look back for first matching opener:
-        opener = closer.previous?
-        opener_found = false
-        while opener && opener != delimiter && opener != openers_bottom[closer_char]
-          odd_match = (closer.can_open || opener.can_close) &&
-                      (opener.orig_delims + closer.orig_delims) % 3 == 0
-          if opener.char == closer.char && opener.can_open && !odd_match
-            opener_found = true
-            break
-          end
-          opener = opener.previous?
-        end
-        opener = nil unless opener_found
-
-        old_closer = closer
-
-        case closer_char
-        when '*', '_'
-          unless opener
+          unless closer.can_close
             closer = closer.next?
-          else
-            # calculate actual number of delimiters used from closer
-            use_delims = (closer.num_delims >= 2 && opener.num_delims >= 2) ? 2 : 1
-            opener_inl = opener.node
-            closer_inl = closer.node
-
-            # remove used delimiters from stack elts and inlines
-            opener.num_delims -= use_delims
-            closer.num_delims -= use_delims
-
-            opener_inl.text = opener_inl.text[0..(-use_delims - 1)]
-            closer_inl.text = closer_inl.text[0..(-use_delims - 1)]
-
-            # build contents for new emph element
-            emph = Node.new((use_delims == 1) ? Node::Type::Emphasis : Node::Type::Strong)
-
-            tmp = opener_inl.next?
-            while tmp && tmp != closer_inl
-              next_node = tmp.next?
-              tmp.unlink
-              emph.append_child(tmp)
-              tmp = next_node
-            end
-
-            opener_inl.insert_after(emph)
-
-            # remove elts between opener and closer in delimiters stack
-            remove_delimiter_between(opener, closer)
-
-            # if opener has 0 delims, remove it and the inline
-            if opener.num_delims == 0
-              opener_inl.unlink
-              remove_delimiter(opener)
-            end
-
-            if closer.num_delims == 0
-              closer_inl.unlink
-              tmp_stack = closer.next?
-              remove_delimiter(closer)
-              closer = tmp_stack
-            end
+            next
           end
-        when '\''
-          closer.node.text = "\u{2019}"
-          if opener
-            opener.node.text = "\u{2018}"
-          end
-          closer = closer.next?
-        when '"'
-          closer.node.text = "\u{201D}"
-          if opener
-            opener.node.text = "\u{201C}"
-          end
-          closer = closer.next?
-        end
 
-        if !opener && !odd_match
-          openers_bottom[closer_char] = old_closer.previous?
-          remove_delimiter(old_closer) if !old_closer.can_open
+          # found emphasis closer. now look back for first matching opener:
+          opener = closer.previous?
+          opener_found = false
+          while opener && opener != delimiter && opener != openers_bottom[closer_char]
+            odd_match = (closer.can_open || opener.can_close) &&
+                        (opener.orig_delims + closer.orig_delims) % 3 == 0
+            if opener.char == closer.char && opener.can_open && !odd_match
+              opener_found = true
+              break
+            end
+            opener = opener.previous?
+          end
+          opener = nil unless opener_found
+
+          old_closer = closer
+
+          case closer_char
+          when '*', '_'
+            unless opener
+              closer = closer.next?
+            else
+              # calculate actual number of delimiters used from closer
+              use_delims = (closer.num_delims >= 2 && opener.num_delims >= 2) ? 2 : 1
+              opener_inl = opener.node
+              closer_inl = closer.node
+
+              # remove used delimiters from stack elts and inlines
+              opener.num_delims -= use_delims
+              closer.num_delims -= use_delims
+
+              opener_inl.text = opener_inl.text[0..(-use_delims - 1)]
+              closer_inl.text = closer_inl.text[0..(-use_delims - 1)]
+
+              # build contents for new emph element
+              emph = Node.new((use_delims == 1) ? Node::Type::Emphasis : Node::Type::Strong)
+
+              tmp = opener_inl.next?
+              while tmp && tmp != closer_inl
+                next_node = tmp.next?
+                tmp.unlink
+                emph.append_child(tmp)
+                tmp = next_node
+              end
+
+              opener_inl.insert_after(emph)
+
+              # remove elts between opener and closer in delimiters stack
+              remove_delimiter_between(opener, closer)
+
+              # if opener has 0 delims, remove it and the inline
+              if opener.num_delims == 0
+                opener_inl.unlink
+                remove_delimiter(opener)
+              end
+
+              if closer.num_delims == 0
+                closer_inl.unlink
+                tmp_stack = closer.next?
+                remove_delimiter(closer)
+                closer = tmp_stack
+              end
+            end
+          when '\''
+            closer.node.text = "\u{2019}"
+            if opener
+              opener.node.text = "\u{2018}"
+            end
+            closer = closer.next?
+          when '"'
+            closer.node.text = "\u{201D}"
+            if opener
+              opener.node.text = "\u{201C}"
+            end
+            closer = closer.next?
+          end
+
+          if !opener && !odd_match
+            openers_bottom[closer_char] = old_closer.previous?
+            remove_delimiter(old_closer) if !old_closer.can_open
+          end
         end
       end