… src/markd/rules/html_block.cr
… src/markd/rules/list.cr
… src/markd/rules/paragraph.cr