@asterite 2019-09-09
Another fact: I tried benchmarking rendering markd's README.md with markd and crystal-cmark, which is just a wrapper around a C library, and these are the results:

```
      markd   5.18k (193.20µs) (± 3.62%)   170kB/op   2.42× slower
common_mark  12.53k ( 79.78µs) (± 4.53%)  6.48kB/op        fastest
```

So markd is pretty good, considering it wasn't written with optimizations in mind (I think there's still room for improvement). Just 2.42 times slower is actually pretty good. The memory consumption there isn't real because the benchmark just shows memory Crystal knows about, not memory consumed by C, so we can't compare that.